pyproject.toml
[tool.poetry]
name = "fileshare"
version = "0.1.0"
description = "A Django App for file sharing across users and teams"
authors = ["Franklin Obasi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.3"
djangorestframework = "^3.15.2"
django-storages = "^1.14.4"
boto3 = "^1.35.66"
django-cors-headers = "^4.6.0"
redis = "^5.2.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
sentry-sdk = {extras = ["django"], version = "^2.19.0"}
flake8 = "^7.1.1"
black = "^24.10.0"
autopep8 = "^2.3.1"
pycodestyle = "^2.12.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
exclude = '''
/(
    \.git
    | \.venv
    | migrations
)/
'''
